`quacc.util` --> `quacc.utils` (#767)
