+++ src/quacc/recipes/emt/core.py
@@ -12,9 +12,9 @@ from ase.optimize import FIRE
 
 from quacc import job
 from quacc.schemas.ase import summarize_opt_run, summarize_run
-from quacc.schemas.atoms import fetch_atoms
-from quacc.util.calc import run_ase_opt, run_calc
-from quacc.util.dicts import merge_dicts
+from quacc.utils.calc import run_ase_opt, run_calc
+from quacc.utils.dicts import merge_dicts
+from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
