+++ src/quacc/recipes/emt/slabs.py
@@ -5,8 +5,8 @@ from typing import TYPE_CHECKING
 
 from quacc import flow, job, subflow
 from quacc.recipes.emt.core import relax_job, static_job
-from quacc.schemas.atoms import fetch_atoms
-from quacc.util.slabs import make_max_slabs_from_bulk
+from quacc.utils.slabs import make_max_slabs_from_bulk
+from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
