+++ src/quacc/recipes/gaussian/core.py
@@ -7,11 +7,11 @@ from typing import TYPE_CHECKING
 from ase.calculators.gaussian import Gaussian
 
 from quacc import job
-from quacc.schemas.atoms import fetch_atoms
 from quacc.schemas.cclib import summarize_run
-from quacc.util.atoms import get_charge, get_multiplicity
-from quacc.util.calc import run_calc
-from quacc.util.dicts import merge_dicts
+from quacc.utils.atoms import get_charge, get_multiplicity
+from quacc.utils.calc import run_calc
+from quacc.utils.dicts import merge_dicts
+from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
