+++ src/quacc/recipes/gulp/core.py
@@ -8,9 +8,9 @@ from ase.calculators.gulp import GULP
 
 from quacc import job
 from quacc.schemas.ase import summarize_run
-from quacc.schemas.atoms import fetch_atoms
-from quacc.util.calc import run_calc
-from quacc.util.dicts import merge_dicts
+from quacc.utils.calc import run_calc
+from quacc.utils.dicts import merge_dicts
+from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
