+++ src/quacc/recipes/lj/core.py
@@ -17,10 +17,10 @@ from quacc.schemas.ase import (
     summarize_thermo_run,
     summarize_vib_run,
 )
-from quacc.schemas.atoms import fetch_atoms
-from quacc.util.calc import run_ase_opt, run_ase_vib, run_calc
-from quacc.util.dicts import merge_dicts
-from quacc.util.thermo import ideal_gas
+from quacc.utils.calc import run_ase_opt, run_ase_vib, run_calc
+from quacc.utils.dicts import merge_dicts
+from quacc.utils.thermo import ideal_gas
+from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
