+++ src/quacc/recipes/orca/core.py
@@ -8,11 +8,11 @@ from typing import TYPE_CHECKING
 from ase.calculators.orca import ORCA, OrcaProfile
 
 from quacc import SETTINGS, job
-from quacc.schemas.atoms import fetch_atoms
 from quacc.schemas.cclib import summarize_run
-from quacc.util.atoms import get_charge, get_multiplicity
-from quacc.util.calc import run_calc
-from quacc.util.dicts import merge_dicts
+from quacc.utils.atoms import get_charge, get_multiplicity
+from quacc.utils.calc import run_calc
+from quacc.utils.dicts import merge_dicts
+from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
