+++ src/quacc/recipes/vasp/qmof.py
@@ -13,10 +13,10 @@ from ase.optimize import BFGSLineSearch
 from quacc import job
 from quacc.calculators.vasp import Vasp
 from quacc.schemas.ase import summarize_opt_run
-from quacc.schemas.atoms import fetch_atoms
 from quacc.schemas.vasp import summarize_run
-from quacc.util.calc import run_ase_opt, run_calc
-from quacc.util.dicts import merge_dicts
+from quacc.utils.calc import run_ase_opt, run_calc
+from quacc.utils.dicts import merge_dicts
+from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
