+++ src/quacc/recipes/vasp/slabs.py
@@ -5,11 +5,11 @@ from typing import TYPE_CHECKING
 
 from quacc import flow, job, subflow
 from quacc.calculators.vasp import Vasp
-from quacc.schemas.atoms import fetch_atoms
 from quacc.schemas.vasp import summarize_run
-from quacc.util.calc import run_calc
-from quacc.util.dicts import merge_dicts
-from quacc.util.slabs import make_adsorbate_structures, make_max_slabs_from_bulk
+from quacc.utils.calc import run_calc
+from quacc.utils.dicts import merge_dicts
+from quacc.utils.slabs import make_adsorbate_structures, make_max_slabs_from_bulk
+from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
