+++ src/quacc/utils/slabs.py
@@ -13,7 +13,7 @@ from pymatgen.analysis.adsorption import AdsorbateSiteFinder
 from pymatgen.core.surface import Slab, center_slab, generate_all_slabs
 from pymatgen.io.ase import AseAtomsAdaptor
 
-from quacc.util.atoms import copy_atoms
+from quacc.utils.atoms import copy_atoms
 
 if TYPE_CHECKING:
     from pymatgen.core import Structure
