+++ src/quacc/utils/wflows.py
@@ -3,9 +3,9 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
+    from ase import Atoms
     from covalent import electron as ct_electron
     from covalent import lattice as ct_lattice
-    from dask_jobqueue.core import DaskJobqueueJob
     from jobflow import Job as JobflowJob
     from parsl.app.python import PythonApp
 
@@ -138,3 +138,33 @@ def subflow(_func: callable | None = None, **kwargs) -> callable | ct_electron:
         raise ValueError(msg)
 
     return decorated
+
+
+def fetch_atoms(atoms: Atoms | dict) -> Atoms:
+    """
+    Returns an Atoms object from a typical quacc recipe input, which can
+    either be an `Atoms` object or a dictionary with an entry `{"atoms": Atoms}`.
+    It may seem a bit silly to do this, but there is a purpose. If you want to
+    write a workflow where the output of one recipe is passed to the input of
+    another recipe, you can always do output["atoms"] to fetch the output Atoms
+    object to pass to the input to the second function. However, this process
+    will often be its own compute step in workflow management tools because they
+    need to resolve the output in order to query it. Depending on the workflow manager,
+    this can be a waste of compute resources, so it's oftentimes better to do this
+    parsing inside the compute task itself, which is why passing in the raw dictionary
+    can be preferred.
+
+    Parameters
+    ----------
+    atoms
+        Atoms object or dictionary with an entry {"atoms": Atoms}
+
+    Returns
+    -------
+    Atoms
+        Atoms object
+    """
+    try:
+        return atoms["atoms"]
+    except Exception:
+        return atoms
