+++ tests/utils/Zn2CuAu.cif.gz
