+++ tests/utils/ZnTe.cif.gz
