+++ tests/utils/test_files.py
@@ -1,6 +1,6 @@
 import os
 
-from quacc.util.files import make_unique_dir
+from quacc.utils.files import make_unique_dir
 
 
 def test_make_unique_dir(tmpdir):
