+++ tests/utils/test_util_atoms.py
@@ -7,7 +7,7 @@ from ase.build import bulk, molecule
 from ase.io import read
 
 from quacc.calculators.vasp import Vasp
-from quacc.util.atoms import check_is_metal, get_atoms_id, prep_next_run
+from quacc.utils.atoms import check_is_metal, get_atoms_id, prep_next_run
 
 FILE_DIR = Path(__file__).resolve().parent
 ATOMS_MAG = read(os.path.join(FILE_DIR, "..", "calculators", "vasp", "OUTCAR_mag.gz"))
